+++ client/src/features/common/LuxOverlay.tsx
@@ -7,12 +7,14 @@ import sanitizeHtml from 'sanitize-html'
 import theme from '../../styles/theme'
 import EntityResultsDescription from '../cms/EntityResultsDescription'
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
+import { pushClientEvent } from '../../lib/pushClientEvent'
 
 const StyledButton = styled(Button)`
   color: ${theme.color.link};
   background-color: white;
   margin-left: 0px;
   text-decoration: none;
+  font-weight: ${theme.font.mobile.bodyLight.weight};
 
   &:hover {
     text-decoration: underline;
@@ -20,6 +22,7 @@ const StyledButton = styled(Button)`
 
   @media (min-width: ${theme.breakpoints.md}px) {
     margin-left: 0.5em;
+    font-weight: ${theme.font.desktop.bodyRegular.weight};
   }
 `
 
@@ -29,11 +32,12 @@ const LuxOverlay: React.FC = () => {
   )
   const [show, setShow] = useState(false)
   const attribRef = useRef(null)
-
+  const overlayText = 'Learn more about Objects and Works'
   const objectsWorksHelpText =
     EntityResultsDescription('objectsWorksHelp') || ''
 
   const onClick: React.MouseEventHandler<HTMLButtonElement> = (): void => {
+    pushClientEvent('Overlay', 'Selected', overlayText)
     setShow(!show)
   }
 
@@ -50,11 +54,11 @@ const LuxOverlay: React.FC = () => {
         variant="link"
         onClick={onClick}
         data-testid="image-attribution-overlay-button"
-        aria-label="Learn more about Objects and Works"
+        aria-label={overlayText}
         className="px-0"
       >
         <i className="bi bi-question-circle open me-1" />
-        Learn more about Objects and Works
+        {overlayText}
       </StyledButton>
       <Modal
         show={show}